+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 1.2.0 -- UNRELEASED
+
+Added new functions `constructor-fn` and `apply-constructor-fns` which can be
+used to decrease verbosity in a configuration.
+
 # 1.1.0 -- 29 Mar 2018
 
 Many functions exposed only for testing have been excluded from